Update CoC guidance
People getting confused about which page to visit to read the CoC that applies to the project.

Just having link to .NET Foundation one which includes the email address to report issues to.

Made sure that page has a big call out at the top of that page giving appropriate kudos to Contributor Covenant for being awesome. But hopefully it's more clear who to email now.